Restructured folders to effectHandler and generalHandler to match the panel's titles
