+++ frontend/src/components/CraftingPanel/CraftingPanel.js
@@ -3,45 +3,14 @@ import './craftingPanel.css';
 import { Button } from "primereact/button";
 import DropdownMenu from "../OptionItemsMenu/DropdownMenu";
 import { ButtonContext } from "../../context/ButtonContext";
-import { TextHandler } from "./craftingComponents/TextHandler/TextHandler";
-import { StylingHandler } from "./craftingComponents/StylingHandler/StylingHandler";
+import { TextHandler } from "./craftingComponents/GeneralHandler/GeneralHandler";
+import { GeneralHandler } from "./craftingComponents/GeneralHandler/GeneralHandler";
+import { EffectHandler } from "./craftingComponents/EffectHandler/EffectHandler";
  
 
 export function CraftingPanel() {
-  const { 
-    buttonGeneralSettings,
-    updateButtonGeneralSettings,
-    buttonTextSettings,
-    updateButtonTextSettings,
-    hoverEffectSettings,
-    updateHoverEffectSettings
-
-  } = useContext(ButtonContext);
 
   const [activeTab, setActiveTab] = useState('General');
-
-  const handleBackgroundColorChange = (e) => {
-    updateButtonGeneralSettings('backgroundColor', e.target.value);
-  };
-  const handleTextColorChange = (e) => {
-    updateButtonTextSettings('textColor', e.target.value);
-  };
-
-  const handleDirectionUp = () => {
-    updateHoverEffectSettings('hoverEffectDirection', 'translateY(-100%)');
-  }
-  const handleDirectionDown = () => {
-    updateHoverEffectSettings('hoverEffectDirection', 'translateY(100%)');
-
-  }
-  const handleDirectionLeft = () => {
-    updateHoverEffectSettings('hoverEffectDirection', 'translateX(-100%)');
-
-  }
-  const handleDirectionRight = () => {
-    updateHoverEffectSettings('hoverEffectDirection', 'translateX(100%)');
-  }
-
 
   return (
     <div className="settings-bar">
@@ -55,75 +24,34 @@ export function CraftingPanel() {
             General
           </button>
           <button
-            className={`tab-button ${activeTab === 'Colors' ? 'active' : ''}`}
-            onClick={() => setActiveTab('Colors')}
+            className={`tab-button ${activeTab === 'Effect' ? 'active' : ''}`}
+            onClick={() => setActiveTab('Effect')}
           >
-            Colors
+            Effect
           </button>
           <button
-            className={`tab-button ${activeTab === 'Effects' ? 'active' : ''}`}
-            onClick={() => setActiveTab('Effects')}
+            className={`tab-button ${activeTab === 'Extra' ? 'active' : ''}`}
+            onClick={() => setActiveTab('Extra')}
           >
-            Effects
+            Extra
           </button>
         </div>
       </div>
       <div className="tab-content">
         {activeTab === 'General' && (
           <div className="settings-options">
-            <TextHandler />
-            <StylingHandler />
+            <GeneralHandler />
           </div>
         )}
-        {activeTab === 'Colors' && (
+        {activeTab === 'Effect' && (
           <div className="settings-options">
-            
-              <div>Text Color</div>
-              <input 
-                type="color"
-                id="TextColor" 
-                value={buttonTextSettings.textColor}
-                onChange={handleTextColorChange}
-                className="color-input"
-                />
-            
             <div className="settings-option">
-              {/* <label htmlFor="bgColor">Secondary Color</label> */}
-              <div>Background Color</div>
-              <input 
-                type="color" 
-                id="bgColor" 
-                value={buttonGeneralSettings.backgroundColor}
-                onChange={handleBackgroundColorChange}
-                />
-            </div>
-            <div className="settings-option">
-              <div className='arrows-buttons-container'>
-                  <div className='arrow-button'>
-                    <Button className="arrow-button-style" icon="pi pi-angle-right" rounded text raised onClick={handleDirectionLeft} />
-                  </div>
-                  <div className='arrow-button'>
-                    <Button className="arrow-button-style" icon="pi pi-angle-left" rounded text size="small" raised onClick={handleDirectionRight}/>
-                  </div>                
-                  <div className='arrow-button'>
-                    <Button className="arrow-button-style" icon="pi pi-angle-down" rounded text raised onClick={handleDirectionUp} />
-                  </div>                
-                  <div className='arrow-button'>
-                    <Button className="arrow-button-style" icon="pi pi-angle-up" rounded text raised onClick={handleDirectionDown} />
-                  </div>
-              </div>
+              <EffectHandler />
             </div>
           </div>
         )}
-        {activeTab === 'Effects' && (
+        {activeTab === 'Extra' && (
           <DropdownMenu />
-          // <div className="settings-options">
-          //   <div className="settings-option">
-          //     {/* <label htmlFor="blurEffect">Blur Effect</label> */}
-          //     <label htmlFor="blurValue">Blur Effect</label>
-          //     <Slider className='slider w-14rem' id='blurValue' value={blurValue} onChange={(e) => setBlurValue(e.value)}  />
-          //   </div>
-          // </div>
         )}
       </div>
     </div>
